made some steps towards xhtml compliancy checked in the W3C DTD's so I can validate documents offline
all documents now declare a namespace and make a document
declaration of "TRANSITIONAL", except in a few cases where
they are bare enough to pass as "Strict".

The next step is to move all layout-specific tags into
stylesheets ( we could probably get away with one or two
for the whole site ) and finish off the rest of the HTML 4.1
compliancy problems there.